Use objc2 for macOS font smoothing determination
`objc2` / `objc2-foundation` is much more type-safe than `cocoa`, and
automatically ensures that memory management-rules are upheld.

See also https://github.com/alacritty/copypasta/pull/74.

+++ src/directwrite/mod.rs
@@ -85,8 +85,8 @@ impl DirectWriteRasterizer {
 
         Ok(RasterizedGlyph {
             character,
-            width: (bounds.right - bounds.left) as i32,
-            height: (bounds.bottom - bounds.top) as i32,
+            width: bounds.right - bounds.left,
+            height: bounds.bottom - bounds.top,
             top: -bounds.top,
             left: bounds.left,
             advance: (0, 0),
